expert-statics page fixed bug
