+++ src/pages/Portal/Header/Header.jsx
@@ -1,6 +1,5 @@
 import React, { useContext, useEffect, useRef, useState } from "react";
 import { Link, useLocation } from "react-router-dom";
-
 import Navbar from "./Navbar";
 import { motion } from "framer-motion";
 import { useDispatch, useSelector } from "react-redux";
@@ -8,11 +7,8 @@ import { languageChange } from "../../../reduxToolkit/languageSlice";
 import i18next from "i18next";
 import { GrayContext } from "../../../component/Layout/Layout";
 import { getContact } from "../../../reduxToolkit/contactSlice/extraReducer";
-
 import "./header.scss";
-
 import Logo from "../../../assets/images/Logo.png";
-import burger from "../../../assets/images/icons/burger.svg";
 import Flag from "../../../assets/images/Flag.png";
 import Blazon from "../../../assets/images/blazon.png";
 import Music from "../../../assets/images/Music.png";
@@ -77,8 +73,7 @@ const HeaderPortal = () => {
               height="15"
               viewBox="0 0 15 15"
               fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
+              xmlns="http://www.w3.org/2000/svg">
               <path
                 fillRule="evenodd"
                 clipRule="evenodd"
@@ -88,8 +83,7 @@ const HeaderPortal = () => {
             </svg>
             <a
               href={`tel: ${contactData?.phone}`}
-              className="header_navbar_phone_number"
-            >
+              className="header_navbar_phone_number">
               {contactData?.phone}
             </a>
           </div>
@@ -99,8 +93,7 @@ const HeaderPortal = () => {
               height="14"
               viewBox="0 0 18 14"
               fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
+              xmlns="http://www.w3.org/2000/svg">
               <path
                 fillRule="evenodd"
                 clipRule="evenodd"
@@ -110,8 +103,7 @@ const HeaderPortal = () => {
             </svg>
             <a
               href={`mailto: ${contactData?.email}`}
-              className="header_navbar_phone_number"
-            >
+              className="header_navbar_phone_number">
               {contactData?.email}
             </a>
           </div>
@@ -151,8 +143,7 @@ const HeaderPortal = () => {
               height="24"
               viewBox="0 0 24 24"
               fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
+              xmlns="http://www.w3.org/2000/svg">
               <circle cx="11.5" cy="11.5" r="7.5" stroke="#F0F0F0" />
               <path
                 d="M21.5 21.5L17 17"
@@ -165,15 +156,13 @@ const HeaderPortal = () => {
           <motion.button
             whileTap={{ scale: 0.6 }}
             className="header_navbar_eye"
-            onClick={() => grayScale()}
-          >
+            onClick={() => grayScale()}>
             <svg
               width="20"
               height="6"
               viewBox="0 0 20 6"
               fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
+              xmlns="http://www.w3.org/2000/svg">
               <path
                 fillRule="evenodd"
                 clipRule="evenodd"
@@ -185,8 +174,7 @@ const HeaderPortal = () => {
           <div className="portal-header-lang">
             <div
               className="portal-header-lang-wrapper"
-              onClick={() => setactiveLangBar((el) => !el)}
-            >
+              onClick={() => setactiveLangBar((el) => !el)}>
               <CiGlobe className="portal-header-lang-icon" />
               <span>
                 {language.split("")[0].toUpperCase() + language.split("")[1]}
@@ -195,38 +183,33 @@ const HeaderPortal = () => {
             </div>
             <div
               className="portal-header-lang-bar"
-              style={activeLangBar ? { display: "flex" } : null}
-            >
+              style={activeLangBar ? { display: "flex" } : null}>
               <p
                 onClick={(e) => {
                   handleChangeLng(e.target.innerText.toLowerCase());
                   setactiveLangBar((el) => !el);
-                }}
-              >
+                }}>
                 Uz
               </p>
               <p
                 onClick={(e) => {
                   handleChangeLng(e.target.innerText.toLowerCase());
                   setactiveLangBar((el) => !el);
-                }}
-              >
+                }}>
                 Ru
               </p>
               <p
                 onClick={(e) => {
                   handleChangeLng(e.target.innerText.toLowerCase());
                   setactiveLangBar((el) => !el);
-                }}
-              >
+                }}>
                 En
               </p>
             </div>
           </div>
           <button
             className="header_navbar_eye burger"
-            onClick={() => setactiveSidebar(!activeSidebar)}
-          >
+            onClick={() => setactiveSidebar(!activeSidebar)}>
             {activeSidebar ? (
               <CgClose className="burger-closeIcon" />
             ) : (
@@ -235,8 +218,7 @@ const HeaderPortal = () => {
                 height="20"
                 viewBox="0 0 20 20"
                 fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
+                xmlns="http://www.w3.org/2000/svg">
                 <path
                   fillRule="evenodd"
                   clipRule="evenodd"
@@ -255,8 +237,7 @@ const HeaderPortal = () => {
 
         <div
           className={activeSidebar ? "overlay overlayActive" : "overlay"}
-          onClick={() => setactiveSidebar(!activeSidebar)}
-        ></div>
+          onClick={() => setactiveSidebar(!activeSidebar)}></div>
       </div>
       <div className="mobile-phone-and-email container">
         <div className="header_navbar_phone">
@@ -265,8 +246,7 @@ const HeaderPortal = () => {
             height="15"
             viewBox="0 0 15 15"
             fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
+            xmlns="http://www.w3.org/2000/svg">
             <path
               fillRule="evenodd"
               clipRule="evenodd"
@@ -276,8 +256,7 @@ const HeaderPortal = () => {
           </svg>
           <a
             href="tel: +998(55)502-22-99"
-            className="header_navbar_phone_number"
-          >
+            className="header_navbar_phone_number">
             +998(55)502-22-99
           </a>
         </div>
@@ -287,8 +266,7 @@ const HeaderPortal = () => {
             height="14"
             viewBox="0 0 18 14"
             fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
+            xmlns="http://www.w3.org/2000/svg">
             <path
               fillRule="evenodd"
               clipRule="evenodd"
@@ -298,8 +276,7 @@ const HeaderPortal = () => {
           </svg>
           <a
             href="mailto: info@vatandoshlarfondi.uz"
-            className="header_navbar_phone_number"
-          >
+            className="header_navbar_phone_number">
             info@vatandoshlarfondi.uz
           </a>
         </div>
