+++ src/pages/Portal/pages/AboutCouncil/components/About/About.jsx
@@ -1,47 +1,13 @@
 import {
   CouncilImage,
-  UserIcon,
-  Globe,
   CopyIcon,
   FacebookIcon,
   TwitterIcon,
   TelegramIcon,
   InstagramIcon,
 } from "../../../../../../assets/images/expert";
+import CouncilStatics from "../../../Expert/components/Council/CouncilStatics";
 import "./About.scss";
-
-const data = [
-  {
-    id: 1,
-    country: "Rossiya",
-    number: 24,
-  },
-  {
-    id: 2,
-    country: "Turkiya",
-    number: 18,
-  },
-  {
-    id: 3,
-    country: "Germaniya",
-    number: 16,
-  },
-  {
-    id: 4,
-    country: "Rossiya",
-    number: 8,
-  },
-  {
-    id: 5,
-    country: "Malayziya",
-    number: 4,
-  },
-  {
-    id: 6,
-    country: "Rossiya",
-    number: 2,
-  },
-];
 
 function About() {
   return (
@@ -110,28 +76,7 @@ function About() {
               </li>
             </ul>
           </div>
-          <div className="about-right">
-            <div>
-              <span>
-                <h5>Ro‘yxatdan o‘tganlar</h5>
-                <img src={UserIcon} />
-              </span>
-              <h4>88</h4>
-              <p>Xalqaro olimlar va ekspertlar</p>
-            </div>
-            <div className="council-bottom">
-              <span>
-                <h5>Xorijiy davlatlar</h5>
-                <img src={Globe} />
-              </span>
-              {data.map((data) => (
-                <span className="council-span">
-                  <h5>{data.country}</h5>
-                  <p>{data.number}</p>
-                </span>
-              ))}
-            </div>
-          </div>
+          <CouncilStatics />
         </div>
       </div>
     </div>
