+++ src/pages/Portal/pages/Expert/components/Council/Council.jsx
@@ -1,43 +1,7 @@
-import {
-  CouncilImage,
-  UserIcon,
-  Globe,
-} from "../../../../../../assets/images/expert";
+import { CouncilImage } from "../../../../../../assets/images/expert";
 import "./Council.scss";
 import { Link } from "react-router-dom";
-
-const data = [
-  {
-    id: 1,
-    country: "Rossiya",
-    number: 24,
-  },
-  {
-    id: 2,
-    country: "Turkiya",
-    number: 18,
-  },
-  {
-    id: 3,
-    country: "Germaniya",
-    number: 16,
-  },
-  {
-    id: 4,
-    country: "Rossiya",
-    number: 8,
-  },
-  {
-    id: 5,
-    country: "Malayziya",
-    number: 4,
-  },
-  {
-    id: 6,
-    country: "Rossiya",
-    number: 2,
-  },
-];
+import CouncilStatics from "./CouncilStatics";
 
 function Council() {
   return (
@@ -60,28 +24,7 @@ function Council() {
             <Link to="/expert/council/about">Batafsil</Link>
           </div>
         </div>
-        <div className="council-right">
-          <div>
-            <span>
-              <h5>Ro‘yxatdan o‘tganlar</h5>
-              <img src={UserIcon} />
-            </span>
-            <h4>88</h4>
-            <p>Xalqaro olimlar va ekspertlar</p>
-          </div>
-          <div className="council-bottom">
-            <span>
-              <h5>Xorijiy davlatlar</h5>
-              <img src={Globe} />
-            </span>
-            {data.map((data) => (
-              <span className="council-span">
-                <h5>{data.country}</h5>
-                <p>{data.number}</p>
-              </span>
-            ))}
-          </div>
-        </div>
+        <CouncilStatics />
       </div>
     </div>
   );
